// src/app/interfaces/report-list-dto.ts
export interface ReportViewGridOptions {
    year?: number;
    TrainingCenterId?: string;
    CampusId?: string;
    DevelopmentRoomId?: string;
    documentNumber?: number;
    documentType?:string;
    name?: string;
    group?:string;
    fEnabled?: boolean;
    page: number;
    pageSize: number;
  }